Fix bug causing some courses not to be returned
Because course information is scraped based on styling, those courses that are styled slightly differently than others would not have their column text scraped properly and thus would be ignored. Specifically, courses whose meeting time cell was a single row (as opposed to multiple) are styled using valign="center" instead of valign="top" and thus were not being found.

This code is still in need of further refactoring as it is a simple copy and paste job changing just the selector.